Created docker-compose.yml for production container

+++ index.js
@@ -17,7 +17,11 @@ const port = config.get('service.port') || 3030;
 const serviceName = config.get('service.name');
 app.listen({ port }, async () => {
     console.log(`${serviceName} service running on port - ${port}`);
-    await sequelize.sync({ force: true, match: /_development$/ });
+    if (process.env.NODE_ENV === 'production') {
+        await sequelize.sync();
+    } else {
+        await sequelize.sync({ force: true, match: /_development$/ });
+    }
     console.log(`${serviceName} database connected!`);
 });
